LMOFreeform: Ensure removal of task after exiting freeform
In some cases the task in floating window remains in recents even
after it is destroyed. This renders a bad looking preview on recents
as well as may trigger a crash in some cases since the task no longer
exists. To avoid this, try to remove the task again after freeform
window is destroyed.

Change-Id: Id6c07c89f7732582b4004417030fe97881b80dd7